task 3.5
Create a new GitHub branch off of the main branch.
Add a login view to your myFlix API, allowing your users to log in to the app. Refer to your API and business logic to make sure you’re asking your users for the right information in the login form.
Add a logout button to your main view so logged in users can log out.
Apply form validation and authentication logic to your login form.
Ensure the login details are stored in localStorage so the login session persists until the user logs out.
Create a signup view for your app. Refer to your API and business logic to make sure you’re asking your users for the right information in the signup form.
Integrate the login and signup view with MainView so they’re displayed as the first screen for non-authenticated users. Upon login, users should see both the list of movies and the button to log out.
Make sure to send the authentication token along with the fetch requests in main-view.jsx to get all movies from your authenticated /movies endpoint on the backend.
Push to your branch, then go to your repository and create the pull request.
Submit a link to your pull request here for your tutor to review.
Once your submission is approved, merge the branch into the main branch.