task 3.7
Directions:

Create a new GitHub branch off of the main branch.
Use React Router to enable routing in your app:
Install react-router and react-router-dom packages by running npm install react-router react-router-dom --save.
Refactor the code in render() method in main-view.jsx to use a BrowserRouter, and add routes for your views.
Replace buttons that use onMovieClick and onBackClick with React Router links. This includes MovieCard, and MovieView in main-view.jsx.
Remove anything related to setSelectedMovie, including the state selectedMovie, since it will no longer be used.
Commit the changes to your branch.
Add a navigation bar at the top of the page.
For unauthenticated users, provide links to the Login/Signup page
For authenticated users, provide links to Home (your MainView), Profile (a view you’ll create in just a moment), and Logout.
Commit the changes to your branch.
Part 2: Create a User Profile View

In this part of the task, you’ll create a new view, a Profile view, so that users can view and alter their information, including their list of favorite movies.

BONUS: SPOTLIGHT ON AI
You're welcome to use AI as a coding assistant to help you with this part of the task, as you learned in the "Spotlight on AI" section of the Exercise.

Directions

Add a user Profile view (linked to the main navigation menu for authenticated users). As per your project brief, this view should:
Display user information. You can obtain the logged in user information by using the /users endpoint and then filtering the returning list by comparing the usernames;
Allow a user to update their user information (username, password, email, date of birth);
Allow a user to deregister;
Display a user's favorite movies as a list. For this you’ll need to filter the movies array with something like let favoriteMovies = movies.filter(m => user.FavoriteMovies.includes(m._id)). With this, favoriteMovies is just a regular array of movie objects which could, for example, be displayed as movie cards (the MovieCard components would need to be passed to the ProfileView component and rendered there);
Add a “Favorite” button to your MovieCard and/or MovieView components, so that logged in users can select a movie to store in their list of favorites;
Allow a user to remove a movie from their list of favorites.
Test that all the links in the navigation bar are working correctly. For example, can a logged in user navigate to their profile?
Apply Bootstrap to your app’s new views and features so that they’re consistent with the rest of your app.
Push to your branch, go to your repository, and create the pull request. Submit a link to your pull request here for your tutor to review.
Once accepted, and before you start the next task, merge the pull request into the main branch.
Bonus: Similar Movies

If you implemented the “similar movies” list in your MovieView back in Exercise 3.4: React Lifestyle Methods, you’ll need to refactor your code. You’ll have to move the logic of producing and rendering "Similar Movies" to inside of movie-view.jsx, because you only have access to the targeted movie ID inside movie-view.jsx.

As a result, you’re going to move the code that renders the “similar movies” list into movie-view.jsx, however, this won't be straightforward (hint: passing function props). Keep in mind that the logic of producing the similarMovies array itself will still remain in main-view.jsx, you’ll just have to pass it as a prop to MovieView. It’s there that the rest of the rendering work will be done.

If you didn’t implement the “similar movies” list in 3.4, it’s not too late. Feel free to add it to your project now in movie-view.jsx!

Don't forget to import MovieCard into movie-view.jsx so that the similar movies feature can use this component!